refactor(outputs): extract MetricCard and ProgressBar helpers
