+++ src/components/outputs/index.ts
@@ -8,6 +8,8 @@ export { BackupCard } from './BackupCard'
 export { CapacityBreakdownList } from './CapacityBreakdownList'
 export { DonutChart, DonutLegend } from './DonutChart'
 export { LonghornCapacityDetails } from './LonghornCapacityDetails'
+export { MetricCard } from './MetricCard'
+export { ProgressBar } from './ProgressBar'
 export { SankeyDiagram } from './SankeyDiagram'
 export { Speedometer } from './Speedometer'
 export { ZfsCapacityDetails } from './ZfsCapacityDetails'
